+++ handright/_core.py
@@ -44,131 +44,73 @@ def handwrite(
         templates = (template,)
     else:
         templates = template
-    pages = _draft(
-        text=text,
-        size=tuple(t.get_background().size for t in templates),
-        top_margin=tuple(t.get_top_margin() for t in templates),
-        bottom_margin=tuple(t.get_bottom_margin() for t in templates),
-        left_margin=tuple(t.get_left_margin() for t in templates),
-        right_margin=tuple(t.get_right_margin() for t in templates),
-        line_spacing=tuple(t.get_line_spacing() for t in templates),
-        font_size=tuple(t.get_font_size() for t in templates),
-        word_spacing=tuple(t.get_word_spacing() for t in templates),
-        line_spacing_sigma=tuple(t.get_line_spacing_sigma() for t in templates),
-        font_size_sigma=tuple(t.get_font_size_sigma() for t in templates),
-        word_spacing_sigma=tuple(t.get_word_spacing_sigma() for t in templates),
-        font=tuple(t.get_font() for t in templates),
-        end_chars=tuple(t.get_end_chars() for t in templates),
-        seed=seed,
-    )
-
-    renderer = _Renderer(
-        background=tuple(t.get_background() for t in templates),
-        fill=tuple(t.get_fill() for t in templates),
-        perturb_x_sigma=tuple(t.get_perturb_x_sigma() for t in templates),
-        perturb_y_sigma=tuple(t.get_perturb_y_sigma() for t in templates),
-        perturb_theta_sigma=tuple(t.get_perturb_theta_sigma()
-                                  for t in templates),
-        seed=seed,
-    )
+    pages = _draft(text, templates, seed)
+    templates = copy_templates(templates)
+    _release_font_resources(templates)
+    renderer = _Renderer(templates, seed)
     return mapper(renderer, pages)
 
 
-def _draft(
-        text: str,
-        size: Sequence[Tuple[int, int]],
-        top_margin: Sequence[int],
-        bottom_margin: Sequence[int],
-        left_margin: Sequence[int],
-        right_margin: Sequence[int],
-        line_spacing: Sequence[int],
-        font_size: Sequence[int],
-        word_spacing: Sequence[int],
-        line_spacing_sigma: Sequence[float],
-        font_size_sigma: Sequence[float],
-        word_spacing_sigma: Sequence[float],
-        font: Sequence,
-        end_chars: Sequence[str],
-        seed: Hashable = None,
-) -> Iterator[Page]:
-    size = itertools.cycle(size)
-    top_margin = itertools.cycle(top_margin)
-    bottom_margin = itertools.cycle(bottom_margin)
-    left_margin = itertools.cycle(left_margin)
-    right_margin = itertools.cycle(right_margin)
-    line_spacing = itertools.cycle(line_spacing)
-    font_size = itertools.cycle(font_size)
-    word_spacing = itertools.cycle(word_spacing)
-    line_spacing_sigma = itertools.cycle(line_spacing_sigma)
-    font_size_sigma = itertools.cycle(font_size_sigma)
-    word_spacing_sigma = itertools.cycle(word_spacing_sigma)
-    font = itertools.cycle(font)
-    end_chars = itertools.cycle(end_chars)
+def _release_font_resources(templates) -> None:
+    for t in templates:
+        t.set_font(None)
 
-    num = itertools.count()
+
+def _draft(text, templates, seed=None) -> Iterator[Page]:
+    template_iter = itertools.cycle(templates)
+    num_iter = itertools.count()
     rand = random.Random(x=seed)
     start = 0
     while start < len(text):
-        page = Page(_INTERNAL_MODE, next(size), _BLACK, next(num))
-        start = _draw_page(
-            page,
-            text,
-            start,
-            top_margin=next(top_margin),
-            bottom_margin=next(bottom_margin),
-            left_margin=next(left_margin),
-            right_margin=next(right_margin),
-            line_spacing=next(line_spacing),
-            font_size=next(font_size),
-            word_spacing=next(word_spacing),
-            line_spacing_sigma=next(line_spacing_sigma),
-            font_size_sigma=next(font_size_sigma),
-            word_spacing_sigma=next(word_spacing_sigma),
-            font=next(font),
-            end_chars=next(end_chars),
-            rand=rand,
-        )
+        template = next(template_iter)
+        page = Page(_INTERNAL_MODE, template.get_size(), _BLACK, next(num_iter))
+        start = _draw_page(page, text, start, template, rand)
         yield page
 
 
-def _draw_page(
-        page: Page,
-        text: str,
-        start: int,
-        top_margin: int,
-        bottom_margin: int,
-        left_margin: int,
-        right_margin: int,
-        line_spacing: int,
-        font_size: int,
-        word_spacing: int,
-        line_spacing_sigma: float,
-        font_size_sigma: float,
-        word_spacing_sigma: float,
-        font,
-        end_chars: str,
-        rand: random.Random,
-) -> int:
-    if page.height() < top_margin + line_spacing + bottom_margin:
+def _check_template(page, template) -> None:
+    if page.height() < (template.get_top_margin() + template.get_line_spacing()
+                        + template.get_bottom_margin()):
         raise LayoutError()
-    if font_size > line_spacing:
+    if template.get_font_size() > template.get_line_spacing():
         raise LayoutError()
-    if page.width() < left_margin + font_size + right_margin:
+    if page.width() < (template.get_left_margin() + template.get_font_size()
+                       + template.get_right_margin()):
         raise LayoutError()
-    if word_spacing <= -font_size // 2:
+    if template.get_word_spacing() <= -template.get_font_size() // 2:
         raise LayoutError()
+
+
+def _draw_page(page, text, start: int, template, rand: random.Random) -> int:
+    _check_template(page, template)
+
+    width = page.width()
+    height = page.height()
+
+    font = template.get_font()
+    top_margin = template.get_top_margin()
+    bottom_margin = template.get_bottom_margin()
+    left_margin = template.get_left_margin()
+    right_margin = template.get_right_margin()
+    line_spacing = template.get_line_spacing()
+    font_size = template.get_font_size()
+    word_spacing = template.get_word_spacing()
+    end_chars = template.get_end_chars()
+    line_spacing_sigma = template.get_line_spacing_sigma()
+    font_size_sigma = template.get_font_size_sigma()
+    word_spacing_sigma = template.get_word_spacing_sigma()
 
     draw = page.draw()
     y = top_margin + line_spacing - font_size
-    while y <= page.height() - bottom_margin - font_size:
-        x = float(left_margin)
+    while y <= height - bottom_margin - font_size:
+        x = left_margin
         while True:
             if text[start] == _NEWLINE:
                 start += 1
                 if start == len(text):
                     return start
                 break
-            if (x > page.width() - right_margin - font_size
+            if (x > width - right_margin - font_size
                     and text[start] not in end_chars):
                 break
             xy = (int(x), int(rand.gauss(y, line_spacing_sigma)))
@@ -198,43 +140,19 @@ class _Renderer(object):
     returning rendered image."""
 
     __slots__ = (
-        "_period",
-        "_background",
-        "_fill",
-        "_perturb_x_sigma",
-        "_perturb_y_sigma",
-        "_perturb_theta_sigma",
+        "_templates",
         "_rand",
         "_hashed_seed",
     )
 
-    def __init__(
-            self,
-            background: Sequence[PIL.Image.Image],
-            fill: Sequence,
-            perturb_x_sigma: Sequence[float],
-            perturb_y_sigma: Sequence[float],
-            perturb_theta_sigma: Sequence[float],
-            seed: Hashable = None,
-    ) -> None:
-        if not (len(background)
-                == len(fill)
-                == len(perturb_x_sigma)
-                == len(perturb_y_sigma)
-                == len(perturb_theta_sigma)):
-            raise ValueError()
-        self._period = len(background)
-        self._background = background
-        self._fill = fill
-        self._perturb_x_sigma = perturb_x_sigma
-        self._perturb_y_sigma = perturb_y_sigma
-        self._perturb_theta_sigma = perturb_theta_sigma
+    def __init__(self, templates, seed=None) -> None:
+        self._templates = templates
         self._rand = random.Random()
         self._hashed_seed = None
         if seed is not None:
             self._hashed_seed = hash(seed)
 
-    def __call__(self, page: Page) -> PIL.Image.Image:
+    def __call__(self, page) -> PIL.Image.Image:
         if self._hashed_seed is None:
             # avoid different processes sharing the same random state
             self._rand.seed()
@@ -242,32 +160,22 @@ class _Renderer(object):
             self._rand.seed(a=self._hashed_seed + page.num)
         return self._perturb_and_merge(page)
 
-    def _perturb_and_merge(self, page: Page) -> PIL.Image.Image:
-        canvas = self._background[page.num % self._period].copy()
+    def _perturb_and_merge(self, page) -> PIL.Image.Image:
+        template = _get_template(self._templates, page.num)
+        canvas = template.get_background().copy()
         bbox = page.image.getbbox()
         if bbox is None:
             return canvas
         strokes = _extract_strokes(page.matrix(), bbox)
-        fill = self._fill[page.num % self._period]
-        x_sigma = self._perturb_x_sigma[page.num % self._period]
-        y_sigma = self._perturb_y_sigma[page.num % self._period]
-        theta_sigma = self._perturb_theta_sigma[page.num % self._period]
-        _draw_strokes(
-            canvas.load(),
-            canvas.size,
-            strokes,
-            fill=fill,
-            x_sigma=x_sigma,
-            y_sigma=y_sigma,
-            theta_sigma=theta_sigma,
-            rand=self._rand,
-        )
+        _draw_strokes(canvas.load(), strokes, template, self._rand)
         return canvas
 
 
-def _extract_strokes(
-        bitmap, bbox: Tuple[int, int, int, int]
-) -> NumericOrderedSet:
+def _get_template(templates, index):
+    return templates[index % len(templates)]
+
+
+def _extract_strokes(bitmap, bbox: Tuple[int, int, int, int]):
     left, upper, right, lower = bbox
     assert left >= 0 and upper >= 0
     # reserve 0xFFFFFFFF as _STROKE_END
@@ -286,10 +194,7 @@ def _extract_strokes(
 
 
 def _extract_stroke(
-        bitmap,
-        start: Tuple[int, int],
-        strokes: NumericOrderedSet,
-        bbox: Tuple[int, int, int, int],
+        bitmap, start: Tuple[int, int], strokes, bbox: Tuple[int, int, int, int]
 ) -> None:
     """Helper function of _extract_strokes() which uses depth first search to
     find the pixels of a glyph."""
@@ -307,16 +212,7 @@ def _extract_stroke(
             stack.append((x + 1, y))
 
 
-def _draw_strokes(
-        bitmap,
-        size: Tuple[int, int],
-        strokes: NumericOrderedSet,
-        fill,
-        x_sigma: float,
-        y_sigma: float,
-        theta_sigma: float,
-        rand: random.Random,
-) -> None:
+def _draw_strokes(bitmap, strokes, template, rand) -> None:
     stroke = []
     min_x = _MAX_INT16_VALUE
     min_y = _MAX_INT16_VALUE
@@ -325,17 +221,7 @@ def _draw_strokes(
     for xy in strokes:
         if xy == _STROKE_END:
             center = ((min_x + max_x) / 2, (min_y + max_y) / 2)
-            _draw_stroke(
-                bitmap,
-                size,
-                stroke,
-                center=center,
-                fill=fill,
-                x_sigma=x_sigma,
-                y_sigma=y_sigma,
-                theta_sigma=theta_sigma,
-                rand=rand,
-            )
+            _draw_stroke(bitmap, stroke, template, center, rand)
             min_x = _MAX_INT16_VALUE
             min_y = _MAX_INT16_VALUE
             max_x = 0
@@ -352,24 +238,21 @@ def _draw_strokes(
 
 def _draw_stroke(
         bitmap,
-        size: Tuple[int, int],
         stroke: Sequence[Tuple[int, int]],
+        template,
         center: Tuple[float, float],
-        fill,
-        x_sigma: float,
-        y_sigma: float,
-        theta_sigma: float,
-        rand: random.Random,
+        rand
 ) -> None:
-    dx = rand.gauss(0, x_sigma)
-    dy = rand.gauss(0, y_sigma)
-    theta = rand.gauss(0, theta_sigma)
+    dx = rand.gauss(0, template.get_perturb_x_sigma())
+    dy = rand.gauss(0, template.get_perturb_y_sigma())
+    theta = rand.gauss(0, template.get_perturb_theta_sigma())
     for x, y in stroke:
         new_x, new_y = _rotate(center, x, y, theta)
         new_x = round(new_x + dx)
         new_y = round(new_y + dy)
-        if 0 <= new_x < size[0] and 0 <= new_y < size[1]:
-            bitmap[new_x, new_y] = fill
+        width, height = template.get_size()
+        if 0 <= new_x < width and 0 <= new_y < height:
+            bitmap[new_x, new_y] = template.get_fill()
 
 
 def _rotate(